hybrid_vdb_deploy_ready/requirements.txt
numpy
sentence-transformers
fastapi
uvicorn
qdrant-client
faiss-cpu
pydantic
